pyproject.toml
[project]
name = "language-learning-assistant"
version = "0.1.0"
description = "AI-powered Speech-to-Speech agent for Vietnamese speakers learning Mandarin Chinese"
readme = "README.md"
license = "MIT"
requires-python = ">=3.12"
authors = [
    { name = "Minh Tang", email = "tangbamiinh@gmail.com" }
]

dependencies = [
    # Hermes Agent (Phase 1 brain)
    "hermes-agent[all]",

    # Qwen3 ASR (Speech-to-Text)
    "qwen-asr>=0.0.6",

    # Qwen3 TTS (Text-to-Speech for Chinese)
    "qwen-tts>=0.1.0",

    # VieNeu-TTS (Vietnamese TTS)
    "vieneu>=1.0.0",

    # Audio utilities
    "soundfile>=0.12.1",
    "numpy>=1.24.0",
    "torch>=2.1.0",

    # HTTP/API bridge
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.34.0",
    "httpx>=0.27.0",
    "python-dotenv>=1.0.0",
    "pydantic>=2.0.0",

    # LLM client (OpenRouter via OpenAI-compatible API)
    "openai>=1.58.0",
]

[project.optional-dependencies]
# GPU acceleration
gpu = [
    "flash-attn>=2.6.0",
]

# vLLM backend for faster ASR inference
vllm = [
    "vllm>=0.6.0",
]

# LiveKit (Phase 2 real-time voice)
# openai plugin is needed for LLM.with_openrouter() — not for STT/TTS
# mlx-audio is needed for local STT on Apple Silicon (Qwen3-ASR-1.7B-8bit)
livekit = [
    "livekit-agents[silero,turn-detector,openai]>=1.4.0",
    "mlx-audio>=0.3.1",
]

# Apple Silicon local inference (MLX backend)
mlx = [
    "mlx-audio>=0.3.1",
]

# Development
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.24.0",
    "pytest-cov>=5.0.0",
    "ruff>=0.8.0",
]

[tool.uv]
dev-dependencies = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.24.0",
    "pytest-cov>=5.0.0",
    "ruff>=0.8.0",
]

[tool.ruff]
target-version = "py312"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP"]
ignore = ["E501"]

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
